backend/requirements.txt
google-generativeai>=0.8.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
pydantic>=2.6.0
pydantic-settings>=2.2.0
faiss-cpu>=1.8.0
numpy>=1.26.0
langchain>=0.2.0
langchain-text-splitters>=0.2.0
fpdf2>=2.7.0
langdetect>=1.0.9
PyPDF2>=3.0.0
aiofiles>=23.2.0
